Move contents of dataviews-patterns to index
